renamed login.go to woBoxForm.go; updated config to release; removed old config.json install template; added release executables
