--- UIMod/config.json
@@ -1,48 +1,0 @@
-{
-  "discordToken": "",
-  "controlChannelID": "",
-  "statusChannelID": "",
-  "connectionListChannelID": "",
-  "logChannelID": "",
-  "saveChannelID": "",
-  "controlPanelChannelID": "",
-  "DiscordCharBufferSize": 0,
-  "blackListFilePath": "",
-  "isDiscordEnabled": false,
-  "errorChannelID": "",
-  "backupKeepLastN": 0,
-  "isCleanupEnabled": false,
-  "backupKeepDailyFor": 0,
-  "backupKeepWeeklyFor": 0,
-  "backupKeepMonthlyFor": 0,
-  "backupCleanupInterval": 0,
-  "backupWaitTime": 0,
-  "gameBranch": "",
-  "ServerName": "Stationeers Server UI",
-  "SaveInfo": "Moon Moon",
-  "ServerMaxPlayers": "6",
-  "ServerPassword": "",
-  "ServerAuthSecret": "",
-  "AdminPassword": "",
-  "GamePort": "27016",
-  "UpdatePort": "27015",
-  "UPNPEnabled": false,
-  "AutoSave": true,
-  "SaveInterval": "300",
-  "AutoPauseServer": true,
-  "LocalIpAddress": "",
-  "StartLocalHost": true,
-  "ServerVisible": true,
-  "UseSteamP2P": true,
-  "ExePath": "",
-  "AdditionalParams": "",
-  "authEnabled": false,
-  "JwtKey": "",
-  "AuthTokenLifetime": 0,
-  "Debug": false,
-  "CreateSSUILogFile": false,
-  "LogLevel": 0,
-  "IsUpdateEnabled": true,
-  "AllowPrereleaseUpdates": false,
-  "AllowMajorUpdates": false
-}
